+++ backend/os/location_test.go
@@ -3,7 +3,6 @@ package os
 import (
 	"os"
 	"path"
-	"path/filepath"
 	"regexp"
 	"testing"
 
@@ -26,7 +25,7 @@ type osLocationTest struct {
 
 func (s *osLocationTest) SetupSuite() {
 	// Directly instantiate the OS FileSystem for the test suite
-	s.fileSystem = &FileSystem{}
+	s.fileSystem = NewFileSystem()
 
 	// Create the temp location using the initialized s.fileSystem
 	var err error
@@ -145,8 +144,7 @@ func (s *osLocationTest) TestChangeDir() {
 func (s *osLocationTest) TestVolume() {
 	volume := s.testFile.Location().Volume()
 
-	// For Unix, this returns an empty string. For windows, it would be something like 'C:'
-	s.Equal(filepath.VolumeName(os.TempDir()), volume)
+	s.Equal("", volume)
 }
 
 func (s *osLocationTest) TestPath() {